refactor: improve project modules
