+++ src/app.module.ts
@@ -3,10 +3,10 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TelegramBotModule } from './telegram-bot/telegram-bot.module';
-import { SharedModule } from './shared/shared.module';
+import { CoreModule } from './core/core.module';
 
 @Module({
-  imports: [TelegramBotModule, SharedModule],
+  imports: [TelegramBotModule, CoreModule],
   controllers: [AppController],
   providers: [AppService],
 })
