+++ src/core/core.module.ts
@@ -1,10 +1,12 @@
-import { Module } from '@nestjs/common';
+import { Module, Global } from '@nestjs/common';
+
 import { NpmStatsService } from './npm-stats/npm-stats.service';
 import { UsersModule } from './users/users.module';
 
+@Global()
 @Module({
   imports: [UsersModule],
   providers: [NpmStatsService],
   exports: [NpmStatsService, UsersModule],
 })
-export class SharedModule {}
+export class CoreModule {}
