+++ src/core/npm-stats/__mocks__/npm-stats.service.ts
