+++ src/core/npm-stats/npm-stats.service.spec.ts
