+++ src/core/npm-stats/npm-stats.service.ts
