+++ src/core/npm-stats/shared/api-npm.model.ts
