+++ src/core/users/shared/models/user.model.ts
