+++ src/telegram-bot/core/bot-core.module.ts
@@ -1,9 +1,10 @@
-import { Module } from '@nestjs/common';
+import { Module, Global } from '@nestjs/common';
 
 import { BotService } from './bot/bot.service';
 
+@Global()
 @Module({
   providers: [BotService],
   exports: [BotService],
 })
-export class BotSharedModule {}
+export class BotCoreModule {}
