+++ src/telegram-bot/core/bot/__mocks__/bot.service.ts
