+++ src/telegram-bot/core/bot/bot.service.spec.ts
