+++ src/telegram-bot/features/add/add.feature.spec.ts
@@ -4,13 +4,13 @@ import { of } from 'rxjs';
 
 import * as TelegramBot from 'node-telegram-bot-api';
 
-import { NpmStatsService } from '../../../shared/npm-stats/npm-stats.service';
-jest.mock('../../../shared/npm-stats/npm-stats.service');
-import { NpmStatsService as NpmStatsServiceMock } from '../../../shared/npm-stats/__mocks__/npm-stats.service';
-import { UsersService } from '../../../shared/users/service/users.service';
-jest.mock('../../../shared/users/service/users.service');
-import { BotService } from '../../shared/bot/bot.service';
-jest.mock('../../shared/bot/bot.service');
+import { NpmStatsService } from '@core/npm-stats/npm-stats.service';
+jest.mock('@core/npm-stats/npm-stats.service');
+import { NpmStatsService as NpmStatsServiceMock } from '@core/npm-stats/__mocks__/npm-stats.service';
+import { UsersService } from '@core/users/service/users.service';
+jest.mock('@core/users/service/users.service');
+import { BotService } from '../../core/bot/bot.service';
+jest.mock('../../core/bot/bot.service');
 
 import { AddFeature } from './add.feature';
