+++ src/telegram-bot/features/add/add.feature.ts
@@ -2,13 +2,13 @@ import { Injectable } from '@nestjs/common';
 
 import * as TelegramBot from 'node-telegram-bot-api';
 
-import { BotService } from '../../shared/bot/bot.service';
+import { BotService } from '../../core/bot/bot.service';
 
 import { BaseCommand } from '../base.command';
 import { Template } from './common';
-import { UsersService } from '../../../shared/users/service/users.service';
+import { UsersService } from '@core/users/service/users.service';
 import { sendMessageHTML } from '../../common';
-import { NpmStatsService } from '../../../shared/npm-stats/npm-stats.service';
+import { NpmStatsService } from '@core/npm-stats/npm-stats.service';
 
 @Injectable()
 export class AddFeature extends BaseCommand {
