+++ src/telegram-bot/features/features.module.ts
@@ -1,14 +1,11 @@
 import { Module } from '@nestjs/common';
 
-import { SharedModule } from '../../shared/shared.module';
-import { BotSharedModule } from '../shared/bot-shared.module';
-
 import { AddFeature } from './add/add.feature';
 import { StartFeature } from './start/start.feature';
 import { StatsModule } from './stats/stats.module';
 
 @Module({
-  imports: [BotSharedModule, SharedModule, StatsModule],
+  imports: [StatsModule],
   providers: [AddFeature, StartFeature],
 })
 export class FeaturesModule {}
