+++ src/telegram-bot/features/stats/stats.feature.ts
@@ -4,10 +4,10 @@ import * as TelegramBot from 'node-telegram-bot-api';
 
 import { BaseCommand } from '../base.command';
 
-import { BotService } from '../../../telegram-bot/shared/bot/bot.service';
+import { BotService } from '../../core/bot/bot.service';
 
-import { NpmStatsService } from '../../../shared/npm-stats/npm-stats.service';
-import { UsersService } from '../../../shared/users/service/users.service';
+import { NpmStatsService } from '@core/npm-stats/npm-stats.service';
+import { UsersService } from '@core/users/service/users.service';
 
 import { sendStatsMsg } from './common';
