+++ src/telegram-bot/telegram-bot.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TelegramBotController } from './telegram-bot.controller';
 import { FeaturesModule } from './features/features.module';
-import { SharedModule } from '../shared/shared.module';
+import { BotCoreModule } from './core/bot-core.module';
 
 @Module({
-  imports: [FeaturesModule, SharedModule],
+  imports: [FeaturesModule, BotCoreModule],
   controllers: [TelegramBotController],
 })
 export class TelegramBotModule {}
