+++ tsconfig.json
@@ -8,7 +8,10 @@
     "target": "es6",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./"
+    "baseUrl": "./",
+    "paths": {
+      "@core/*": ["src/core/*"]
+    }
   },
   "exclude": ["node_modules"]
 }
